requirements.txt
numpy==1.12.1
psycopg2==2.7.6.1
RPi.GPIO==0.6.5